weather_api_solution/requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.32
Flask-Migrate==4.0.7
flasgger==0.9.7.1
python-dateutil==2.9.0.post0
pytest==8.2.2
